Crate bigbang

Galaxy creator for Space Settler

This program creates a new galaxy for Space Settler. It has a statistically correct sample of stars, and in those stars it creates complete solar systems (still in development). It creates stars regarding to proportion shown here. Then, the planets are created thinking on habitability, but following as far as possible all the physical laws.

Modules

consts
Constants

planet
Planet module

star
Star module

utils
Utilities for Space Settler Galaxy Creator

Structs

Stats