Module bigbang::consts

Constants

This module contains all the needed constants in the galaxy creation.

Constants

AU
The astronomical unit (au or AU), the average distance from earth to the sun, in meters (m).

BOLTZ
σ, the Stefan–Boltzmann constant, in W·m⁻²·K⁻⁴.

C
c, the speed of light in vacuum, in m·s⁻¹.

CH_LIMIT
The Chandrasekhar limit, in kg.

EARTH_ATM_PRESSURE
The atmospheric pressure at sea level in Earth, in pascals (Pa).

EARTH_MASS
The mass of the Earth (M⊕), in kg.

EARTH_RADIUS
The radius of the Earth (R⊕), in meters (m).

G
G, the gravitational constant, in m³·kg⁻¹·s⁻².

JUPITER_MASS
The mass of Jupiter (Mj), in kg.

JUPITER_RADIUS
The radius of Jupiter (Rj), in meters (m).

SUN_LUMINOSITY
The luminosity of the Sun (L☉), in watts (W).

SUN_MASS
The mass of the Sun (M☉), in kg.

SUN_RADIUS
The radius of the Sun (R☉), in meters (m).