Struct bigbang::planet::Atmosphere

pub struct Atmosphere {
    // some fields omitted
}

Atmosphere structure

This structure defines an atmosphere with all the needed parameters and chemical elements. This will be used not only as a mere curiosity but to calculate greenhouse effect and habitability of the body.

Methods

impl Atmosphere

fn get_pressure(&self) -> f64

Get pressure

Gets the pressure of the atmosphere in Pascals (Pa).

fn get_h2o(&self) -> f64

Get H₂O

Gets the percentage of H₂O (water vapour) in the atmosphere, from 0 to 1.

fn get_co2(&self) -> f64

Get CO₂

Gets the percentage of CO₂ (carbon dioxide) in the atmosphere, from 0 to 1.

fn get_co(&self) -> f64

Get CO

Gets the percentage of CO (carbon monoxide) in the atmosphere, from 0 to 1.

fn get_n2(&self) -> f64

Get N₂

Gets the percentage of N₂ (nitrogen) in the atmosphere, from 0 to 1.

fn get_o2(&self) -> f64

Get O₂

Gets the percentage of O₂ (oxygen) in the atmosphere, from 0 to 1.

fn get_ar(&self) -> f64

Get Ar

Gets the percentage of Ar (argon) in the atmosphere, from 0 to 1.

fn get_so2(&self) -> f64

Get SO₂

Gets the percentage of SO₂ (sulfur dioxide) in the atmosphere, from 0 to 1.

fn get_ne(&self) -> f64

Get Ne

Gets the percentage of Ne (neon) in the atmosphere, from 0 to 1.

fn get_ch4(&self) -> f64

Get CH₄

Gets the percentage of CH₄ (methane) in the atmosphere, from 0 to 1.

fn get_he(&self) -> f64

Get He

Gets the percentage of He (helium) in the atmosphere, from 0 to 1.