Struct bigbang::planet::Surface

pub struct Surface {
    // some fields omitted
}

Surface structure

This structure represents the composition of the surface of the planet. Gaseous planets will not have it, while rocky for ones it will contain the surface composition in 3 variables (water, snow and land) in percentages.

Methods

impl Surface

fn get_fresh_water(&self) -> f64

Get fresh water

Gets the percentage of fresh water on the surface of the planet. From 0 to 1.

fn get_ocean_water(&self) -> f64

Get ocean water

Gets the percentage of ocean water on the surface of the planet. From 0 to 1.

fn get_snow(&self) -> f64

Get snow

Gets the percentage of the surface of the planet covered in snow. From 0 to 1.

fn get_land(&self) -> f64

Get land

Gets the percentage of land on the surface of the planet. From 0 to 1.