Enum bigbang::star::StarClass

pub enum StarClass {
    BlackHole,
    NeutronStar,
    QuarkStar,
    WhiteDwarf,
    O,
    B,
    A,
    F,
    G,
    K,
    M,
}

StarClass enumeration

This enumeration contains all the star types available in this universe creation.

Variants

BlackHole

This represents a black hole, an structure so dense that not even the light can escape. Due to the high radiation, no planets can survive near one of these structures. Their mass is between 2 and 20 solar masses. Approximately the 0.03333% of the stars in the galaxy will be of this type. Its effective temperature will be represented as 0K.

NeutronStar

This represents a neutron star. Neutron stars are really compact stars, that due to their extreme gravitational forces not even atomic nuclei can survive, and the protons are degraded to neutrons. Their mass is between 1.1 and 2.2 solar masses. Due to the high radiation, no planet can survive near a neutron star. Some neutron stars can be seen as pulsars. Approximately the 0.03333% of the stars in the galaxy will be of this type.

QuarkStar

This represents a quark star. Quark stars are hypothetical stars that are even more compact than neutron stars but they have not collapsed into black holes. They have between 2 and 3 solar masses. They have such big pressures that even the quarks that are part of the neutrons get free. Approximately the 0.00001% of the stars in the galaxy will be of this type, so most of the galaxies will not even have one.

WhiteDwarf

This represents a white dwarf. A white dwarf is a compact star that is usually the remnant of a Sun-type star. They are very dense and hot, and their size is inversely proportional to the cube of their mass, which can be between 0.2 and 1.39 solar masses. This last mass is the Chandrasekhar limit. There are not many planets around quark stars, and never more than 2. Approximately the 0.03332% of the stars in the galaxy will be of this type.

O

This represents an O type star. This stars are hot and blue colored. Their surface temperature can be as high as 52,000K. Approximately the 0.00003% of the stars will be of this type, so many galaxies will not even have one. No planet can survive to the high radiation coming from the star.

B

This represents a B type star. This stars are hot and blue-white colored. Their surface temperature is usually between 10,000K and 30,000K and their mass between 2.1 and 16 solar masses. Approximately the 0.13% of the stars will be of this type. No more than 2 planets will survive to the radiations, and only in small B type stars.

A

This represents an A type of star. This stars are white stars. Their surface temperature is usually between 7,500K and 10,000K and their mass between 1.4 and 2.1 solar masses. Approximately the 0.6% of the stars will be of this type. They will have up to 3 planets orbiting them.

F

This represents an F type of star. This stars are yellow-white stars. Their surface temperature is usually between 6,000K and 7,500K and their mass between 1.04 and 1.4 solar masses. Approximately the 3% of the stars will be of this type. They will have up to 8 planets orbiting them.

G

This represents an G type of star, the same type as our Sun. This stars are yellow stars. Their surface temperature is usually between 5,200K and 6,000K and their mass between 0.8 and 1.04 solar masses. Approximately the 7.6% of the stars will be of this type. They will have up to 10 planets orbiting them.

K

This represents an K type of star. This stars are orange stars. Their surface temperature is usually between 3,700K and 5,200K and their mass between 0.45 and 0.8 solar masses. Approximately the 12.1% of the stars will be of this type. They will have up to 10 planets orbiting them.

M

This represents an M type of star. This stars are red stars. Their surface temperature is usually between 2,400K and 3,700K and their mass between 0.08 and 0.45 solar masses. Approximately the 76.45% of the stars will be of this type. They will have up to 9 planets orbiting them.